+++ src/engine/index.ts
@@ -1,0 +1,4 @@
+export { OrbitCamera } from './OrbitCamera';
+export { Ray } from './Ray';
+export { GizmoRenderer } from './GizmoRenderer';
+export type { GizmoMode } from './GizmoRenderer';